+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Minimum java version is now java-21, please upgrade if you want to continue building/running LINSTOR.
 - DRBD: create-md now only times out if there was no IO for 45s
 - mkfs: mkfs.ext4 and mkfs.xfs now only time out if there was no IO for 45s
+- Moved old property "ExtCmdWaitTimeout" -> "ExtCmd/WaitTimeout"
 
 ### Removed